Added some clearance, and strengthened the clamp.
For final, I'm thinking lattice glass with this system.  It doesn't
actually need the removable metal surface, but it's a nice add.

Not sure if the thickness is correct/good for the glass, but will
find out.  The glass is cheaper and lighter than the aluminum, by
my checking out of the specs at least.