Added a direct-drive BMG mount
Based on Sean's design, but integrated the belt mount and motor
mount.

Moved fan to the back, which I think is John's design, cuz I want
to make a more focused axial fan, and that's where there's room.

Plan to mount bltouch on the front - maybe on fan, or extend the
mount somehow.