Made a flange for mounting the fabric / mid-seal.
current design: fans up top in front, blowing down through the
headband and right on the face-part of the shield.

Cloth going over the forehead/velcro'd to it and the upper mask rim
to seal up the top fan intake.

Cloth also covering the middle hole in the sheild - it needs to be
flexible for comfort, but mostly sealed so that the fan doesn't
recirculate much.

For exhale, a cloth bib attached to edge of shield and running
under the chin - elastic through a slit in the chin fabric to connect
up to the ears, right where the head-attach elastic sits.

Head almost entirely exposed.  Fans are 5v, run on USB power banks
kept in pockets, off-shield.