Added File to repository
Adding simplified config_bundle. Adds a list of known printers to an allow list for the Nova-labs Print Settings and Filament settings. This is to exclude them from the XL and mini with Input shaping enabled. 

This is not the correct bundle to use if you have a DIY printer, but if you only ever use "modern" printers this should be great.

Also fold the profile for HunkaHunka into the bundle.